+++ src/main/java/org/javaswift/joss/headers/object/range/ExcludeStartRange.java
@@ -10,12 +10,12 @@ public class ExcludeStartRange extends AbstractRange {
     }
 
     @Override
-    public int getFrom(int byteArrayLength) {
+    public long getFrom(int byteArrayLength) {
         return this.offset;
     }
 
     @Override
-    public int getTo(int byteArrayLength) {
+    public long getTo(int byteArrayLength) {
         return byteArrayLength;
     }
 }
