+++ src/main/java/org/javaswift/joss/headers/object/range/LastPartRange.java
@@ -10,12 +10,12 @@ public class LastPartRange extends AbstractRange {
     }
 
     @Override
-    public int getFrom(int byteArrayLength) {
+    public long getFrom(int byteArrayLength) {
         return byteArrayLength - length;
     }
 
     @Override
-    public int getTo(int byteArrayLength) {
+    public long getTo(int byteArrayLength) {
         return byteArrayLength;
     }
 }
